+++ docs/CHANGELOG.md
@@ -17,6 +17,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Update application confirm deadline
+- Update travel page text
+- Update global styles for `a` tag
 
 ### Fixed